pyproject.toml
[tool.poetry]
name = "pdftoolkit"
version = "0.1.0"
description = ""
authors = ["zimolab <zimolab@aliyun.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11, <3.12"
pyguiadapter = "^1.0.6"
pymupdf = "^1.24.13"
pillow = "^11.0.0"
py-multitasking = "^0.1.1"

python-i18n = "^0.3.9"
platformdirs = "^4.3.6"
pyqtdarktheme = "^2.1.0"
loguru = "^0.7.3"
[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
priority = "primary"